✨ Add basic usage example and implement syntax highlighting for Markdown
